refactor: 명칭 수정
